requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
django-ninja==1.4.1
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
